refactor : 패키지 및 클래스명과 구조 변경
