+++ src/main/java/wooteco/subway/fare/domain/farebyagestrategy/DiscountFareByAgeStrategy.java
@@ -1,0 +1,8 @@
+package wooteco.subway.fare.domain.farebyagestrategy;
+
+public interface DiscountFareByAgeStrategy {
+
+    boolean isInAgeRange(int age);
+
+    int discountFareByAge(int fare);
+}
