+++ src/main/java/wooteco/subway/fare/domain/farebydistancestrategy/AdditionalFareOver10Km.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.domain.CalculateAdditionalFare;
+package wooteco.subway.fare.domain.farebydistancestrategy;
 
 public class AdditionalFareOver10Km implements AdditionalFareStrategy {
