+++ src/main/java/wooteco/subway/fare/domain/farebydistancestrategy/AdditionalFareUnder10km.java
@@ -1,4 +1,4 @@
-package wooteco.subway.path.domain.CalculateAdditionalFare;
+package wooteco.subway.fare.domain.farebydistancestrategy;
 
 public class AdditionalFareUnder10km implements AdditionalFareStrategy {
